fix ElectronicProductData and done UI, logic of ElectronicProductForm

+++ Products/ClothingProduct.cs
@@ -8,7 +8,7 @@ namespace OOP_finalProject.Products
     {
         private string size;
         public string Size { get { return size; } set { size = value; } }
-        public ClothingProduct(string id, string name, decimal price, int quantity, string size) : base(id, name, price, quantity)
+        public ClothingProduct(string id, string name, decimal price, decimal quantity, string size) : base(id, name, price, quantity)
         {
             Size = size;
         }
